pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atomdisc"
version = "0.1.0"
description = "AtomDisc: Molecule-aware LoRA fine-tuning toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "AtomDisc Team" }
]
dependencies = [
  "torch==2.3.1",
  "torchvision==0.18.1",
  "torchaudio==2.3.1",
  "torch-scatter>=2.1.2",
  "torch-sparse>=0.6.18",
  "torch-geometric>=2.5.3",
  "rdkit-pypi==2022.9.5",
  "ogb==1.2.0",
  "transformers==4.44.2",
  "tokenizers==0.19.1",
  "accelerate==1.2.1",
  "peft==0.15.2",
  "vector-quantize-pytorch==1.17.1",
  "einops==0.8.0",
  "einx==0.3.0",
  "datasets==2.18.0",
  "evaluate==0.4.3",
  "numpy==1.25.2",
  "pandas==2.2.3",
  "scikit-learn==1.7.0",
  "tqdm==4.66.5",
  "sentencepiece==0.2.0",
  "safetensors==0.5.3",
  "PyYAML==6.0.1"
]

[project.optional-dependencies]
dev = [
  "black==24.10.0",
  "ruff==0.11.0",
  "pytest==7.4.4"
]

[tool.setuptools.packages.find]
include = ["atomdisc*"]


